ai/pyproject.toml
[project]
name = "impetusai-ai"
version = "0.1.0"
description = "ImpetusAI Workplace Platform — AI Agents and Orchestration"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",
    "litellm>=1.40.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",
]

[tool.setuptools.packages.find]
where = [".."]
include = ["ai", "ai.*"]
